feat(website): add `app` dir (#8869)
