+++ apps/website/src/app/docs/[...slug]/layout.tsx
@@ -1,0 +1,14 @@
+import type { PropsWithChildren } from 'react';
+import { Providers } from './providers';
+import { CmdKDialog } from '~/components/CmdK';
+import { Header } from '~/components/Header';
+
+export default function SidebarLayout({ children }: PropsWithChildren<any>) {
+	return (
+		<Providers>
+			<Header />
+			<>{children}</>
+			<CmdKDialog />
+		</Providers>
+	);
+}
