+++ apps/website/src/components/DocContainer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type {
 	ApiItemJSON,
 	TokenDocumentation,
