+++ apps/website/src/components/HyperlinkedText.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { TokenDocumentation } from '@discordjs/api-extractor-utils';
 import Link from 'next/link';
 
@@ -11,7 +13,6 @@ export function HyperlinkedText({ tokens }: { tokens: TokenDocumentation[] }) {
 							className="text-blurple focus:ring-width-2 focus:ring-blurple rounded outline-0 focus:ring"
 							href={token.path}
 							key={idx}
-							prefetch={false}
 						>
 							{token.text}
 						</Link>
