+++ apps/website/src/components/InheritanceText.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { InheritanceData } from '@discordjs/api-extractor-utils';
 import Link from 'next/link';
 
@@ -8,7 +10,6 @@ export function InheritanceText({ data }: { data: InheritanceData }) {
 			<Link
 				className="text-blurple focus:ring-width-2 focus:ring-blurple rounded font-mono outline-0 focus:ring"
 				href={data.path}
-				prefetch={false}
 			>
 				{data.parentName}
 			</Link>
