+++ apps/website/src/components/MDXRemote.tsx
@@ -1,0 +1,3 @@
+'use client';
+
+export { MDXRemote } from 'next-mdx-remote';
