+++ apps/website/src/components/MethodItem.tsx
@@ -1,5 +1,7 @@
+'use client';
+
 import type { ApiMethodJSON, ApiMethodSignatureJSON } from '@discordjs/api-extractor-utils';
-import { Menu, MenuButton, MenuItem, useMenuState } from 'ariakit';
+import { Menu, MenuButton, MenuItem, useMenuState } from 'ariakit/menu';
 import { useCallback, useMemo, useState } from 'react';
 import { FiLink } from 'react-icons/fi';
 import { VscChevronDown, VscVersions } from 'react-icons/vsc';
