+++ apps/website/src/components/ParameterTable.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { ParameterDocumentation } from '@discordjs/api-extractor-utils';
 import { useMemo } from 'react';
 import { HyperlinkedText } from './HyperlinkedText';
