+++ apps/website/src/components/Sections.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type {
 	ApiClassJSON,
 	ApiInterfaceJSON,
