+++ apps/website/src/components/Table.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useMemo, type ReactNode } from 'react';
 
 export function Table({
