+++ apps/website/src/components/model/Interface.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { ApiInterfaceJSON } from '@discordjs/api-extractor-utils';
 import { DocContainer } from '../DocContainer';
 import { MethodsSection, PropertiesSection } from '../Sections';
