+++ apps/website/src/components/model/TypeAlias.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { ApiTypeAliasJSON } from '@discordjs/api-extractor-utils';
 import { DocContainer } from '../DocContainer';
