+++ apps/website/src/contexts/cmdK.tsx
@@ -1,6 +1,8 @@
-import { type DisclosureState, useDialogState } from 'ariakit';
-import type { PropsWithChildren } from 'react';
-import { createContext, useContext } from 'react';
+'use client';
+
+import { useDialogState } from 'ariakit/dialog';
+import type { DisclosureState } from 'ariakit/disclosure';
+import { type PropsWithChildren, createContext, useContext } from 'react';
 
 export const CmdKContext = createContext<DisclosureState | null>(null);
