+++ apps/website/src/styles/main.css
@@ -23,6 +23,10 @@ body {
 	height: 100%;
 }
 
+[data-nextjs-scroll-focus-boundary] {
+	height: 100%;
+}
+
 [data-theme='dark'] {
 	display: none;
 }
